moved fake nuke module
