+++ src/nuke_api/__init__.py
@@ -1,0 +1,6 @@
+try:
+    # if this succeeds, it means that we are running inside Nuke
+    import nuke
+except ImportError:
+    # otherwise import the fake nuke api
+    from . import _nuke_fake_api as nuke
